requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
frozenlist==1.2.0
idna==3.3
loguru==0.5.3
multidict==5.2.0
python-dotenv==0.19.2
pytz==2021.3
typing_extensions==4.0.0
win32-setctime==1.0.3
yarl==1.7.2
